ESP32-C3 RC

+++ ESP-IDF_Robot/main/blink_example_main.c
@@ -804,7 +804,7 @@ void onDataReceived (uint8_t *mac_addr, uint8_t *data, uint8_t data_len) {
     ESP_LOGW(TAG, "Data was received");
     ESP_LOGI(TAG, "x-axis: 0x%04X", (uint8_t)buf->x_axis);
     ESP_LOGI(TAG, "y-axis: 0x%04X", (uint8_t)buf->y_axis);
-    ESP_LOGI(TAG, "PCM 1: 0x%04X", (uint8_t)buf->motor1_rpm_pcm);
+    ESP_LOGI(TAG, "PCM 1: 0x%04X", buf->motor1_rpm_pcm);
     
 }
